[OMPT] Fix abort on nested synchronization regions
Similar to the work callback, there are certain barriers which are able
to be nested (for example taskgroup and taskwait). Right now, this
causes the measurement to abort since we only allow a single barrier in
our tasks. With this commit, nested barriers are allowed.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>
(cherry picked from commit 646494d62d5f1db4104f568fb86b24a3031728b5)