OMPT: Support for OpenMP teams
This commit adds support for OpenMP teams to Score-P. In general, teams
are handeled just like parallel regions right now, with slight
differences to account for the fact that teams might not use an implicit
barrier at the end, even though it is a parallel region.

The checks to leagues were removed from all callbacks and are replaced
by a warning for now.

Also added an open issue found in LLVM based compilers which can occur
when using single nowait or masked within the teams construct.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>