libwrap: Give GOTCHA the chance to wrap a function later
GOTCHA might not find the original function in the current list of
libraries and will return `GOTCHA_FUNCTION_NOT_FOUND` when enabling the
wrapper. Though GOTCHA will search for the unmapped originals every time
a new library is loaded, thus mapping the original function later. As this
mapping is GOTCHA-internal, the documented way of accessing the original
function pointer is by calling `gotcha_get_wrappee` for every wrapped
function call.