OMPT: Add OpenMP device tracing interface
This is the main part of the device tracing interface, adding support
for OpenMP target directives.

The commit is split into two parts, the host side callbacks and the
device tracing interface for accelerator events. We register new
callbacks being used for writing host events while also gathering some
information for accelerator record evaluation.

When a device is initialized, we try to initialize the device tracing
interface, giving us access to accelerator events. If this fails, no
events will be recorded for this accelerator.

The whole OpenMP target support is split into multiple files
representing logical grous based on the callbacks they are based on.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>