OMPT: Add remediable check for OpenMP teams
Added a remediable check for teams with LLVM.
LLVM compilers dispatch non-conforming callbacks when initinalizing
teams. We see an additional implicit task callback which introduces a
nesting layer not actually being there in the application. Since this
later introduces issues with Scalsaca, we want to remove this layer.

Right now, this workaround will be used for all runtimes with OMPT
support, since they are based on LLVM, which includes this bug. This may
change in future compiler releases.

More details about this issue can be seen in the LLVM repo issue:
https://github.com/llvm/llvm-project/issues/78425

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>