Init: Allow independent init. of confvars
Since some adapters might initialize extremely early during program
startup, where we do not want to initialize Score-P already (for example
OMPT when the OpenMP target directive is used), we intend to postpone
the initialization to a later point (see !398). However, this prevents
usage of confvars for the affected adapters. This commit allows those
adapters to initialize the confvars only without initializing the
whole measurement system.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>