Deprecate --openmp and fix issues with OMPT
The option `--openmp` is around for quite a while, but is showing
issues with installations where OMPT is the default, being unable to
find OPARI2 headers.

To fix the issues, `--openmp` will now point towards the OPARI2 adapter,
which matches the current documentation. In addition, deprecate the
option for future Score-P releases, as OMPT will (at some point in the
future) be the only available option and is the default for OpenMP
already.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>