OMPT: Add registration checks to configure
With Cray compilers registering some callbacks with
`ompt_set_sometimes`, we need to make sure that no issues occur
because of that.

This commit is the first step towards handling runtimes not reporting
`ompt_set_always`. Callbacks where the runtime requires
`ompt_set_always` are now explicitly checked. In addition, we require
`ompt_callback_sync_region` to report either `ompt_set_sometimes_paired`
or `ompt_set_always` to enable the OMPT adapter.

In addition, the critical and remediable checks have been modified
slightly. Now, the checks return an exit code of 0 when a callback cannot be
registered, since we cannot reliably test the callback and disable it
anyway during runtime. With this, the configure summary will not get
cluttered with failed checks which may not actually be runtime issues.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>