Re-enable `-tcollect` for compiler instrumentation
Classic Intel compilers were using `-finstrument-functions` instead of
`-tcollect` due to the ordering of the compiler instrumentation check.
This was done intentionally for Score-P v8.0. However, since we lose
compile-time filtering for the classic Intel compilers by doing this,
re-enable this option by moving the check.

To allow configure to succeed, also add a new option to the Intel
platform files to ignore a remark when the required environment
variables are not set and update the documentation to only mention the
classic Intel compilers for full automated compiler instrumentation.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>