addr2line: Ignore ld-linux[x], include end
We have encountered issues with Classic Intel compilers and dlopened
libraries, where the address space was included in ld-linux-x86-64.so.
This causes lookups to fail. Since ld-linux is a linker invoked during
runtime and will likely not point to any actual source lines, ignore it
to resolve runtime loaded libraries correctly.

Also include the end of a library when checking for addresses. This
might fix corner cases, where the searched address is exactly at the
end, though not encountered yet.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>