Allow additional flags for external dependencies
For some distribution packages like libbfd on OpenSUSE, additional flags
are required to use the libraries. Previously, this would result in
configure aborting due to compile errors. With the new
[x]_EXTRA_{CPPFLAGS|LDFLAGS|LIBS}, one can now pass these additional
required flags so that configure passes normally. This resolves issues
when building RPM packages, where using the download option was not
possible.

Co-authored-by: Bert Wesarg <bert.wesarg@tu-dresden.de>
Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>