Instrumenter: Explicitly pre-process after OPARI2
The Cray programming environment, or more specifically, the Cray
compiler environment handles preprocessing differently than we would
normally expect. When an accelerator architecture is loaded, building
the pre-processed file with added code by OPARI2 fails, since the compiler
gets confused by the duplicate code for host and accelerator architectures.
This could be solved by adding certain flags, but is very specific to Cray
compiler versions and not portable between wrappers.

The more portable solution is to explicitly invoke the pre-processing
and writing the resulting file to an extension which tells the compiler
that this file is pre-processed. This works around the Cray issue and also
works for other compilers.

For this to work, we need to add another explicit pre-process step to our
pipeline. This is needed since compilers like GCC strictly refuse to
pre-process something marked as already pre-processed.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>