OPEN_ISSUES: Add OpenMP entries for CPE
With an accelerator architecture loaded, the OMPT checks might fail
when being executed, as the Cray runtime expects that accelerators
can be initialized. This fails when the node doesn't have any GPUs.
Therefore, add an OPEN_ISSUES entry describing that Score-P has to
be built on a GPU node if OMPT support is desired.

In addition, CPE 23.09 and lower will not be able to build any OpenMP
program with OPARI2, due to multiple definitions and other compilation
errors. These are introduced by the compiler wrappers and are unreliable
and complicated to work around. This issue is fixed with CPE 23.12.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>