OMPT: Hashmap if fast transfer is not available
Using target_data->value to transfer our data only works if sizeof(
void* ) <= sizeof( ompt_id_t ). If not, trying to set pointers will
have side-effects, most likely a segmentation fault, when accessing them
on a `buffer-complete` event. To circumvent this, use a hash map for
this exact scenario. This will have a significantly higher overhead,
but is unavoidable if we still want to offer device tracing.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>