OMPT: Sort completed buffers before writing
The OpenMP spec. does not require buffers returned by the runtime to be
sorted. We just experience this with the AMD compilers. To ensure that
other runtimes will not cause issues, sort buffers before writing them
to our locations.
This may reduce the number of created locations.

By counting the number of events on the host-side, we are also able to
remove the reliance on `ompt_callback_target_emi` for device tracing
buffers. This is important, as some runtimes do not implement, or
prevent registration, of this callback for device tracing buffers
e.g. Cray compilers.
This would cause the previous implementation to leak memory.
The current implementation expects that implementations of the device
tracing interface support `ompt_record_target_kernel_t`
and `ompt_record_target_data_op_t` when the tracing interface can be
enabled, and the host-side callbacks can be enabled.
This is the case for all implementations we are aware of.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>