[OMPT] Fix abort on recording event toggle
When a user stops recording Score-P events before a parallel region and
resumes it after the parallel region, the OMPT adapter would abort the
measurement, since it tries to write exit events for the overdue events.
This would cause the measurement to abort.

To fix this, store if the user had recording events enabled when
entering a parallel region. Use this value for all places where overdue
events would write an exit event.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>