LLVM plug-in: Handle exceptions after optimization
When handling exceptions at the start of the optimization
pipeline, possible inlined functions may be missed. This introduces a
lot of overhead compared to an un-instrumented run. In the case of
LULESH, exception handling caused a ~8x overhead.

This can be improved by moving this pass to the end of the pipeline.
Here, all functions are inlined already. With this, the overhead drops
significantly to around 1.7x.

This change will not affect builds without inlined functions (e.g. the
Debug build with CMake).

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>