addr2line: Make SoLookup functions thread-safe
The functions `SCOREP_Addr2line_SoLookupAddr()` and
`SCOREP_Addr2line_SoLookupAddrRange()` use a provided `soHandle` to
lookup an address. We need to ensure that the `soHandle` is still
valid before passing it to libbfd. A handle can become invalid if it
belongs to a runtime-loaded shared object that was dlclosed.