CUDA: Inject NVTX to avoid using LD_PRELOAD
With NVTX v2 and older, NVTX has to be enabled by LD_PRELOADing
the CUDA adapter when executing the instrumented application.
With NVTX v3, the event library can be injected via
NVTX_INJECTION64_PATH or setting a weak symbol.
To allow this, implement both options. During configure,
the build type is decided by trying to find NVTX v3 first.
If not found, the old approach is reused.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>