CUDA: Implement NVTX version check
Since CUDA 10.0, two NVTX versions are delivered with CUDA.
NVTX v1/v2 are deprecated versions, which rely on a shared-library
to be linked. NVTX v3 is header-only. In addition, v3 allows
to inject a tool easily by either setting a weak symbol or
setting some environment variables.

This commit adds a configure check to detect which NVTX version
we're dealing with. If v3 is found, it is preferred. If not,
we fall back to v2. If NVTX is not found at all, the respective
part of the CUDA adapter is disabled.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>