Conf.: Generate include help for empty header list
AFS_EXTERNAL_LIB generates help entries automatically, based on the
variables passed to the call. However, the include option is only
generated when one passes at least one header to this call. However,
there are scenarios where multiple headers might be checked, where not
all of them are available. In this case, no header should be passed, but
the help option should still show up. Therefore, change the help output
generation so the include help is generated even when an empty string is
passed for the header.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>