LLVM plug-in: Use opaque pointer for global string
LLVM has started to use opaque pointers for a long time. However,
starting with LLVM 17, these are the default with typed pointers not
being available anymore.

Since LLVM 20 deprecates the old CreateGlobalStringPtr function,
pointing to CreateGlobalString, and ROCm 6.0.3 on LUMI crashes when
using this function, transition to CreateGlobalString beginning with
LLVM 17.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>