Build: Fix abort on "yes"/"no" substring for libs
AFS_EXTERNAL_LIB checks if "yes"/"no" is set for
`--with-lib<library>-include` and/or `--with-lib<library>-lib`.
However, that check was too broad, checking for any matching substring
in the given path. This could fail if some folder included "yes" or
"no". Make the check more strict to ensure that we do only abort when
"yes" or "no" is actually passed

Co-authored-by: Bert Wesarg <bert.wesarg@tu-dresden.de>
Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>