OMPT: Fix incorrect reduction codeptr_ra
The previous implementation assumed that at most one reduction happens
at the end of a construct accepting the reduction clause. However,
runtimes are allowed to split the reduction into multiple steps, where
only some of the threads actually doing the reduction are active. Since
we assumed that only one reduction would happen, we would reset the
codeptr_ra. The second reduction would then incorrectly point to the
outer parallel region.

To fix this, use the fact that an implementation barrier surrounds the
reduction operation in (almost) all cases. With this, we get both a more
accurate codeptr_ra and can reuse this value for multiple reductions.

This fails for reductions which do not run with multiple threads. This
was seen for a loop construct outside of any parallel region. However,
this is an edge case where a reduction is questionable anyways.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>