LLVM plug-in: Improve filter strictness
Remove `_omp_` from filtering rules, as this was filtering user
functions as well. Change some rules to check with `strncmp` to make
sure this doesn't happen there as well.

Additionally, allow users to overwrite our filter rules by checking if
the user INCLUDES a function in his filter file. If so, enable
instrumentation for this function.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>