OMPT: Ingore internal implicit tasks for teams
This commit adds a workaround for teams with LLVM.
LLVM compilers dispatch non-conforming callbacks when initinalizing
teams. We see an additional implicit task callback which introduces a
nesting layer not actually being there in the application. Since this
later introduces issues with Scalsaca, detecting nesting where there is
actually no nesting due to additional recorded events, we want to
remove this layer.

To achieve this, we use a boolean value to note that a parallel region
was created by a league. We can use this information, together with the
encountering task, to filter this region in `parallel_begin`,
`parallel_end` and `implicit_task`.

To ensure that this workaround is only used for runtimes showing this
bug, the last commit did add a macro we're using here to remove the
workaround for runtimes adhering to the specifications.

Right now, this workaround will be used for all runtimes with OMPT
support, since they are based on LLVM, which includes this bug. This may
change in future compiler releases.

More details about this issue can be seen in the LLVM repo issue:
https://github.com/llvm/llvm-project/issues/78425

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>