OMPT: Disable reduction callback
Since handling the reduction callback works on the Score-P side, but can
produce issues with handling them in Scalasca due to similar behavior to
OpenMP tasks, and also can create significant overhead due to the very
simple nature (e.g. '+') for most reductions, this feature should be
optional.

As this is not yet possible, disable the callback until confvars are
implemented.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>