Profiling: Move found node to head of sibling list
When profiling nodes have many siblings (e.g., GPU kernels that are
created as children of an artificial root node), the linear search
to find an existing call path on entering a region can become a
bottleneck.  Assuming that most applications operate in phases, move
the identified node to the head of the sibling list, thus keeping
the most recently used nodes near the head of the list.