Build: Ignore 'llc' flag for Fortran LDFLAGS
The macros AC_F77_LIBRARY_LDFLAGS and AC_FC_LIBRARY_LDFLAGS
try to retrieve required linker flags for a conftest so that
this program can build correctly. For this, the verbose output
of the compile and link command is parsed. Here, arguments
with `-L` and `-l` are extracted.

This typically worked and `configure` succeeded, but was
error-prone. On an `aarch64` system with NVHPC 24.9 however,
this flag extraction caused the flag `-lsr-complexity-limit`,
passed to `llc` during the compile process` to be picked up
as well. As this is not a linker flag, all further tests using
`FLIBS` or `FCLIBS` would fail. This lead to an abort of
`configure`.

To fix this, ignore this flag by removing it from
`FLIBS` and `FCLIBS`. Since this issue might also occur with
other Clang-based compilers, this removal does not depend on
the platform.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>