Make scorep-score self-contained
scorep-score's SOURCES have a direct dependency to libcube4 which
needs to be stated.

scorep_score_DEPENDENCIES was removed. It contained just values that
are already listed in scorep_score_LDADD. The latter LDADD are used to
compute the former DEPENDENCIES (if this variable isn't present). It
is unclear why DEPENDENCIES was used in the first place.