MPIF08, adapter: add F08 interface
The interface is used by the F08 MPI wrappers.
Interfaces to functions in adapter/mpi/management are in the same
directory. The Fortran side is in files *_h.F90, the C side is in files
*_f08.c.
Interfaces to functions in other parts of Score-P are in scorep_f08 with
the same naming convention.