LLVM plugin: Do not instrument when not available
In rare cases, e.g. when mixing HPE CPE CCE and ROCm, the configure
check may run, but fail after writing the file consumed by
build-backend. This would cause the LLVM plugin to not build, but still
tried to be used when instrumenting something.
To fix this, explicitly delete the file when the compile test in
the build-llvm-plugin configure failed.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>