MPIF08, adapter: track requests only if event gen is on
Resolves issue #434.

This is to prevent a bug with MPICH and F08: A call to a routine with
choice buffers from 'use mpi_f08' enters our F08 wrapper and then also
our C wrapper. In the case of MPI_Irecv (and some other calls) this will
lead to doubly tracked requests (and ultimately program crashes). This
commit prevents request tracking when the `scorep_mpi_generate_events`
flag is off, i.e., when a wrapper was entered from another wrapper.