CUDA: Fix incorrect location count without memcopy
When an instrumented code does not contain any memcopy operations, or
they were not detected by our CUDA adapter, e.g. in the case of async
operations, the location count was still incremented when determining
the amount of locations. This caused the unification to be ran with an
invalid interim communicator handle, causing memory corruption of
`allocator->allocated_memory`. This then caused a segmentation fault
when trying to free this pointer during cleanup.

To fix this, check if CUDA contexts have a valid location id before
incrementing the location count, as is already done for streams.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>