OMPT: Ensure initialized device before target event
While the OpenMP standard theoretically allows having the device
initialization nested within the target-begin event, as this event is
just before starting execution of the target construct on the
accelerator device, it causes issues for us, since we want to
differentiate between the device having device tracing or not.
Previously, the target-begin event would just return, not creating a
target_data entry for the target-end event (and all in between). During
the target-end event, we run into a segmentation fault due to the
missing target_data.

Working around this is possible, but may have side effects, e.g. still
getting events in the trace buffer when we handle events as not having
device tracing, or leaking memory since the data is not freed correctly.

Since the issue is quite rare, mostly happening when the first target
region is empty, or copying scalar variables with default mapping
on the stack, abort the measurement instead pointing to this being a
runtime deficiency.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>